Move get_all_voters fn
